Left Rotation
A  left rotation operation on a circular array shifts each of the array's elements  unit to the left. The elements that fall off the left end reappear at the right end. Given an integer , rotate the array that many steps to the left and return the result.

Example
d=2
arr=[1,2,3,4]
After  2 rotations, arr=[3,4,5,1,2] 

Function Description

Complete the  function with the following parameters:

: the amount to rotate by
: the array to rotate
Returns

: the rotated array