Print the Elements of a Linked List
This exercise focuses on traversing a linked list. You are given a pointer to the head node of a linked list. The task is to print the data of each node, one per line. If the head pointer is null , indicating the list is empty, nothing should be printed.

Function Description

Complete the printlinkedlist function with the following parameter(s):

: a reference to the head of the list
Print

For each node, print its  value on a new line (console.log in Javascript).
Input Format

The first line of input contains , the number of elements in the linked list.
The next  lines contain one element each, the  values for each node.

Note: Do not read any input from stdin/console. Complete the printLinkedList function in the editor below.