Create Tree: Inorder Traversal
Complete the  function in your editor below, which has  parameter: a pointer to the root of a binary tree. It must print the values in the tree's inorder traversal as a single line of space-separated values.

Input Format

Our hidden tester code passes the root node of a binary tree to your $inOrder* function.

Constraints


Output Format

Print the tree's inorder traversal as a single line of space-separated values.