Sparse Arrays
There is a collection of input strings and a collection of query strings. For each query string, determine how many times it occurs in the list of input strings. Return an array of the results.

Example



There are  instances of '',  of '', and  of ''. For each query, add an element to the return array: .

Function Description

Complete the function  with the following parameters:

: an array of strings to search
: an array of query strings
Returns
int [q]: the results of each query